Correct variable usage tracking within array literals
Also fix some style unconformities.

+++ src/runtime.cc
@@ -300,16 +300,22 @@ Obj ArrayToString(Array *arr)
 	intptr_t len = arr->length ? getInteger(arr->length) : 0;
 	bool first = true;
 	std::string str = "[";
-	for (int i = 0; i < len; ++i) {
+	for (intptr_t i = 0; i < len; ++i)
+	{
 		std::string value = "null";
-		if (arr->buffer[i]) {
-			if (!isInteger(arr->buffer[i]) && arr->buffer[i]->isa == &StringClass) {
+		if (arr->buffer[i])
+		{
+			if (!isInteger(arr->buffer[i]) && arr->buffer[i]->isa == &StringClass)
+			{
 				value = "\"" + std::string(reinterpret_cast<String *>(arr->buffer[i])->characters) + "\"";
-			} else {
+			}
+			else
+			{
 				CompiledMethod mth = compiledMethodForSelector(arr->buffer[i], sel);
 				assert(mth);
 				Obj returned = callCompiledMethod(mth, arr->buffer[i], sel, nullptr, 0);
-				if (returned->isa == &StringClass) {
+				if (returned->isa == &StringClass)
+				{
 					value = reinterpret_cast<String *>(returned)->characters;
 				}
 			}
